refactoring: change location dir
