+++ src/core/locale/LanguageManager.cpp
