+++ src/core/locale/LanguageManager.h
